// src/interfaces/user-goal.ts
export interface UserGoal {
    uid?: string;
    name?: string;
    completed_percent?: number;
    start_date?: string;
    status?: string;
    goal?: string;
    user?: string;
    description?: string;
    picture?: string;
    category?: string;
    company?: string;
    timestamp?: string;
    percent?:string;
	subcategory?:string;
}